Safari-611.1.10/LayoutTests/fast/animation/request-animation-frame-missing-arguments.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test how animation methods react to too few arguments.');
shouldThrow("requestAnimationFrame()");
shouldThrow("cancelAnimationFrame()");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
